@@ -4,16 +4,16 @@ description: A Flutter starter app that gives you a solid foundation for a real 
 # The following defines the version and build number for your application.
 # A version number is three numbers separated by dots, like 1.2.43
 # followed by an optional build number separated by a +.
-# Both the version and the builder number may be overridden in flutter
+# Both the version and the build number may be overridden in flutter
 # build by specifying --build-name and --build-number, respectively.
 # Read more about versioning at semver.org.
-version: 1.6.9+27
+version: 1.6.10+28
 
 # Prevent accidental publishing to pub.dev.
 publish_to: "none"
 
 environment:
-  sdk: ">=3.0.0 <4.0.0"
+  sdk: ">=3.2.0 <4.0.0"
 
 dependencies:
   flutter:
@@ -23,18 +23,18 @@ dependencies:
   intl: ^0.18.0
   oauth2_client: ^3.0.0
   cupertino_icons: ^1.0.5
-  meta: ^1.7.0
-  http: ^0.13.5
+  meta: ^1.10.0
+  http: ^1.1.0
   provider: ^6.0.3
   shared_preferences: ^2.0.15
   google_maps_flutter: ^2.2.1
-  google_maps_flutter_web: ^0.4.0+3
+  google_maps_flutter_web: ^0.5.4+2
   geocoding: ^2.0.5
   firebase_core: ^2.4.0
   firebase_analytics: ^10.0.8
   firebase_crashlytics: ^3.0.8
   firebase_messaging: ^14.2.0
-  geolocator: ^9.0.1
+  geolocator: ^10.1.0
   mockito: ^5.3.2
 
 dev_dependencies:
@@ -42,8 +42,8 @@ dev_dependencies:
     sdk: flutter
   integration_test:
     sdk: flutter
-  flutter_lints: ^2.0.1
-  test: ^1.21.4
+  flutter_lints: ^3.0.1
+  test: ^1.24.3
   flutter_launcher_icons: ^0.13.1
   convert: ^3.0.2
 
