@@ -11,7 +11,7 @@ import 'package:first_app/ui/pages/map/index.dart';
 import 'drawer.dart';
 
 class HomePage extends StatelessWidget {
-  const HomePage({Key? key}) : super(key: key);
+  const HomePage({super.key});
 
   @override
   Widget build(BuildContext context) {
