@@ -1,5 +1,5 @@
 // File generated by FlutterFire CLI.
-// ignore_for_file: lines_longer_than_80_chars
+// ignore_for_file: lines_longer_than_80_chars, avoid_classes_with_only_static_members
 import 'package:firebase_core/firebase_core.dart' show FirebaseOptions;
 import 'package:flutter/foundation.dart'
     show defaultTargetPlatform, kIsWeb, TargetPlatform;
@@ -19,18 +19,29 @@ class DefaultFirebaseOptions {
     if (kIsWeb) {
       return web;
     }
-    // ignore: missing_enum_constant_in_switch
     switch (defaultTargetPlatform) {
       case TargetPlatform.android:
         return android;
       case TargetPlatform.iOS:
         return ios;
       case TargetPlatform.macOS:
-        return macos;
+        throw UnsupportedError(
+          'DefaultFirebaseOptions have not been configured for macos - '
+          'you can reconfigure this by running the FlutterFire CLI again.',
+        );
+      case TargetPlatform.windows:
+        throw UnsupportedError(
+          'DefaultFirebaseOptions have not been configured for windows - '
+          'you can reconfigure this by running the FlutterFire CLI again.',
+        );
+      case TargetPlatform.linux:
+        throw UnsupportedError(
+          'DefaultFirebaseOptions have not been configured for linux - '
+          'you can reconfigure this by running the FlutterFire CLI again.',
+        );
       default:
         throw UnsupportedError(
-          'Failed to find a default FirebaseOptions for the current platform: $defaultTargetPlatform.\n'
-          'Please ensure that you have imported the correct Firebase SDK for this platform.',
+          'DefaultFirebaseOptions are not supported for this platform.',
         );
     }
   }
@@ -62,20 +73,7 @@ class DefaultFirebaseOptions {
     projectId: 'actingweb-firstapp',
     databaseURL: 'https://actingweb-firstapp.firebaseio.com',
     storageBucket: 'actingweb-firstapp.appspot.com',
-    iosClientId:
-        '748007732162-20hnc1u6qmn35avunaidj44p43o6g4s7.apps.googleusercontent.com',
-    iosBundleId: 'io.actingweb.firstapp',
-  );
-
-  static const FirebaseOptions macos = FirebaseOptions(
-    apiKey: 'AIzaSyCFGeVdzEh6ZPKClV_0in0xHHX-0NtqGQ8',
-    appId: '1:748007732162:ios:b662cf09a8209a8b785c5b',
-    messagingSenderId: '748007732162',
-    projectId: 'actingweb-firstapp',
-    databaseURL: 'https://actingweb-firstapp.firebaseio.com',
-    storageBucket: 'actingweb-firstapp.appspot.com',
-    iosClientId:
-        '748007732162-20hnc1u6qmn35avunaidj44p43o6g4s7.apps.googleusercontent.com',
+    iosClientId: '748007732162-20hnc1u6qmn35avunaidj44p43o6g4s7.apps.googleusercontent.com',
     iosBundleId: 'io.actingweb.firstapp',
   );
 }
