@@ -8,11 +8,11 @@ class AnchoredOverlay extends StatelessWidget {
   final Widget? child;
 
   const AnchoredOverlay({
-    Key? key,
+    super.key,
     this.showOverlay,
     this.overlayBuilder,
     this.child,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -38,11 +38,11 @@ class OverlayBuilder extends StatefulWidget {
   final Widget? child;
 
   const OverlayBuilder({
-    Key? key,
+    super.key,
     this.showOverlay = false,
     this.overlayBuilder,
     this.child,
-  }) : super(key: key);
+  });
 
   @override
   OverlayBuilderState createState() => OverlayBuilderState();
@@ -118,10 +118,10 @@ class CenterAbout extends StatelessWidget {
   final Widget? child;
 
   const CenterAbout({
-    Key? key,
+    super.key,
     this.position,
     this.child,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
